feat: calculate distributed and remain cashback

+++ internal/dto/statistic_dto.go
@@ -1,13 +1,13 @@
 package dto
 
 type Cashback struct {
-	Distributed float32 `json:"distributed"`
-	Remain      float32 `json:"remain"`
+	Distributed float64 `json:"distributed"`
+	Remain      float64 `json:"remain"`
 }
 
 type StatisticSummaryResponse struct {
 	NumOfOrders      int      `json:"num_of_orders"`
 	NumOfCustomers   int      `json:"num_of_customers"`
-	TotalRevenue     float32  `json:"total_revenue"`
+	TotalRevenue     float64  `json:"total_revenue"`
 	TotalASACashback Cashback `json:"total_asa_cashback"`
 }
